agrego cambios server Proyecto Final

+++ TP3/Proyecto_Integrador/eliminar_linea_bd.py
@@ -1,8 +1,9 @@
 from modules.config import app, db
-from modules.databases import Reclamo_db
+from modules.databases import Persona_db, Reclamo_db
 
 with app.app_context():
     db.create_all()
-    reclamo_a_eliminar = db.session.query(Reclamo_db).filter_by(ID_reclamo=x).first()
-    db.session.delete(reclamo_a_eliminar)
-    db.session.commit()
+    reclamo_a_eliminar = db.session.query(Reclamo_db).filter_by(ID_reclamo=0).first()
+    if reclamo_a_eliminar is not None:
+        db.session.delete(reclamo_a_eliminar)
+        db.session.commit()
